requirements.txt
########################################################################
# # BASE FOR FASTAPI
fastapi==0.95.2
uvicorn[standard]==0.21.1
pydantic==1.*
########################################################################
# # BASE FOR WORKER
python-dotenv
celery==5.3.1
flower==1.0.0
redis==3.5.3
Cython==0.29.21
pydub==0.25.1
gevent
boto3==1.26.148
########################################################################
# # Sadtalker
numpy==1.23.4
face_alignment==1.3.5
imageio==2.19.3
imageio-ffmpeg==0.4.7
librosa==0.9.2 # 
numba
resampy==0.3.1
pydub==0.25.1 
scipy==1.10.1
kornia==0.6.8
tqdm
yacs==0.1.8
pyyaml  
joblib==1.1.0
scikit-image==0.19.3
basicsr==1.4.2
facexlib==0.3.0
gradio
gfpgan
av
safetensors
moviepy==1.0.3
realesrgan